Update README.md
In this code, the user is prompted to enter a number using the input() function, and the input is converted to a floating-point number using the float() function.

The square of the number is then calculated by raising it to the power of 2 using the ** operator. The result is stored in the variable square.

Finally, the result is printed to the console using the print() function.